Доработки
Добавили ещё несколько классов для реализации возможностей hibernate.validator'а, но, к сожалению, h.v всё ещё остаётся без дела. Поправили несколько комментариев при прочтении методички.

+++ src/main/java/ru/gb/antonov/j611/beans/repos/RolesRepo.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import ru.gb.antonov.j611.entities.Role;
 
 @Repository
-public interface RolesRepo extends CrudRepository <Role, Long>  {}
+public interface RolesRepo extends CrudRepository <Role, Integer>  {}
